feat: create app departament and all routes
