+++ _core/settings.py
@@ -53,6 +53,7 @@ MY_APPS = [
     "employees",
     "suppliers",
     "invoices",
+    "departament"
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + MY_APPS
