+++ _core/urls.py
@@ -19,5 +19,5 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/", include("employees.urls")),
-    path("api/", include("suppliers.urls")),
+    path("api/", include("departament.urls")),
 ]
