+++ suppliers/views.py
@@ -1,17 +1,3 @@
-from rest_framework import generics
-from .serializer import DepartamentSerializer
-from .models import Departament
-from django.shortcuts import get_object_or_404
+from django.shortcuts import render
 
-class DepartamentView(generics.ListCreateAPIView):
-
-
-    serializer_class = DepartamentSerializer
-    queryset = Departament.objects.all()
-
-
-class DepartamentDetailView(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Departament.objects.all()
-    serializer_class = DepartamentSerializer
-    
-    lookup_url_kwarg = "pk"
+# Create your views here.
